+++ assembly/instructions/store.ts
@@ -34,13 +34,13 @@ export const store_imm_u64: InstructionRun = (r, args, _registers, memory) => {
 
 // STORE_U8
 export const store_u8: InstructionRun = (r, args, registers, memory) => {
-  memory.setU8(faultRes, args.b, u8(registers[reg(args.a)] & 0xff));
+  memory.setU8(faultRes, args.b, u8(registers[reg(args.a)] & u64(0xff)));
   return okOrFault(r, faultRes);
 };
 
 // STORE_U16
 export const store_u16: InstructionRun = (r, args, registers, memory) => {
-  memory.setU16(faultRes, args.b, u16(registers[reg(args.a)] & 0xff_ff));
+  memory.setU16(faultRes, args.b, u16(registers[reg(args.a)] & u64(0xff_ff)));
   return okOrFault(r, faultRes);
 };
 
@@ -87,14 +87,14 @@ export const store_imm_ind_u64: InstructionRun = (r, args, registers, memory) =>
 // STORE_IND_U8
 export const store_ind_u8: InstructionRun = (r, args, registers, memory) => {
   const address = u32(registers[reg(args.a)] + u32SignExtend(args.c));
-  memory.setU8(faultRes, address, u8(registers[reg(args.b)] & 0xff));
+  memory.setU8(faultRes, address, u8(registers[reg(args.b)] & u64(0xff)));
   return okOrFault(r, faultRes);
 };
 
 // STORE_IND_U16
 export const store_ind_u16: InstructionRun = (r, args, registers, memory) => {
   const address = u32(registers[reg(args.a)] + u32SignExtend(args.c));
-  memory.setU16(faultRes, address, u16(registers[reg(args.b)] & 0xff_ff));
+  memory.setU16(faultRes, address, u16(registers[reg(args.b)] & u64(0xff_ff)));
   return okOrFault(r, faultRes);
 };
